(USBHDFSD) Changed sceUsbdOpenPipeAligned() to sceUsbdOpenPipe(), removed code that breaks up long transfers (since they're too long to exceed on the IOP).
The change to sceUsbdOpenPipe() is to avoid triggering the driver's checks when the address is misaligned.

+++ iop/usb/usbhdfsd/src/include/mass_stor.h
@@ -5,8 +5,8 @@ int InitUSB(void);
 int mass_stor_disconnect(int devId);
 int mass_stor_connect(int devId);
 int mass_stor_probe(int devId);
-int mass_stor_readSector(mass_dev* dev, unsigned int sector, unsigned char* buffer, unsigned int count);
-int mass_stor_writeSector(mass_dev* dev, unsigned int sector, const unsigned char* buffer, unsigned int count);
+int mass_stor_readSector(mass_dev* dev, unsigned int sector, unsigned char* buffer, unsigned short int count);
+int mass_stor_writeSector(mass_dev* dev, unsigned int sector, const unsigned char* buffer, unsigned short int count);
 int mass_stor_configureNextDevice(void);
 
 #endif
